+++ src/detectorator.h
@@ -31,21 +31,21 @@ namespace detectorator_namespace
 	{
 	private:
 		cv::Mat img;
-		void resizeImg(double);
+		void resizeImg(float);
 		void reportFailedOperation(std::string, bool);
 		void performAnOperationWithPixels(PixelsOperation);
-		double binaryThreshValue;
-		double gaussConst; /// Constant subtracted from the mean of weighted mean
-		double imgCompressPercentage;
+		float binaryThreshValue;
+		float gaussConst; /// Constant subtracted from the mean of weighted mean
+		float imgCompressPercentage;
 		int gaussBlockSize; /// Size of a pixel neighbourhood : 3, 5, 7, ...
 		bool isPixelMatchesPatterns(vInt &, PixelPatterns &, cf::CommonFunctions &);
 	public:
 		void execute();
 		void readImg(fs::path);
 		void writeImg(fs::path);
-		void setBinaryThreshValue(double value) { binaryThreshValue = value; };
-		void setImgCompressPercentage(double value) { imgCompressPercentage = value; };
-		void setGaussConst(double value) { gaussConst = value; };
+		void setBinaryThreshValue(float value) { binaryThreshValue = value; };
+		void setImgCompressPercentage(float value) { imgCompressPercentage = value; };
+		void setGaussConst(float value) { gaussConst = value; };
 		void setGaussBlockSize(int value) { gaussBlockSize = value; };
 	};
 }
